[chore]: People Profile component folders organized
