+++ src/components/People/PeopleProfile/CopyBtn/CopyBtn.tsx
